+++ openfisca_core/errors/__init__.py
@@ -21,6 +21,7 @@
 #
 # See: https://www.python.org/dev/peps/pep-0008/#imports
 
+from .empty_argument_error import EmptyArgumentError  # noqa: F401
 from .parameter_not_found_error import ParameterNotFoundError, ParameterNotFoundError as ParameterNotFound  # noqa: F401
 from .parameter_parsing_error import ParameterParsingError  # noqa: F401
 from .period_mismatch_error import PeriodMismatchError  # noqa: F401
