+++ openfisca_core/taxscales/abstract_tax_scale.py
@@ -1,0 +1,45 @@
+from __future__ import annotations
+
+import typing
+import warnings
+
+import numpy
+
+from openfisca_core import taxscales
+
+
+class AbstractTaxScale(taxscales.TaxScaleLike):
+    """
+    Base class for various types of tax scales: amount-based tax scales, rate-based
+    tax scales...
+    """
+
+    def __init__(self, name: typing.Optional[str] = None, option = None, unit = None) -> None:
+        message = [
+            "The 'AbstractTaxScale' class has been deprecated since version 34.7.0,",
+            "and will be removed in the future.",
+            ]
+        warnings.warn(" ".join(message), DeprecationWarning)
+        super(AbstractTaxScale, self).__init__(name, option, unit)
+
+    def __repr__(self) -> typing.NoReturn:
+        raise NotImplementedError(
+            "Method '__repr__' is not implemented for "
+            f"{self.__class__.__name__}",
+            )
+
+    def calc(
+            self,
+            tax_base: typing.Union[numpy.ndarray[int], numpy.ndarray[float]],
+            right: bool,
+            ) -> typing.NoReturn:
+        raise NotImplementedError(
+            "Method 'calc' is not implemented for "
+            f"{self.__class__.__name__}",
+            )
+
+    def to_dict(self) -> typing.NoReturn:
+        raise NotImplementedError(
+            f"Method 'to_dict' is not implemented for "
+            f"{self.__class__.__name__}",
+            )
