+++ openfisca_core/taxscales/helpers.py
@@ -1,0 +1,39 @@
+import logging
+import typing
+
+from openfisca_core import parameters, taxscales
+
+log = logging.getLogger(__name__)
+
+
+def combine_tax_scales(
+        node: parameters.ParameterNodeAtInstant,  # node: ParameterNodeAtInstant
+        combined_tax_scales: typing.Optional[taxscales.MarginalRateTaxScale] = None,
+        ) -> typing.Optional[taxscales.MarginalRateTaxScale]:
+    """
+    Combine all the MarginalRateTaxScales in the node into a single
+    MarginalRateTaxScale.
+    """
+
+    name = next(iter(node or []), None)
+
+    if name is None:
+        return combined_tax_scales
+
+    if combined_tax_scales is None:
+        combined_tax_scales = taxscales.MarginalRateTaxScale(name = name)
+        combined_tax_scales.add_bracket(0, 0)
+
+    for child_name in node:
+        child = node[child_name]
+
+        if isinstance(child, taxscales.MarginalRateTaxScale):
+            combined_tax_scales.add_tax_scale(child)
+
+        else:
+            log.info(
+                f"Skipping {child_name} with value {child} "
+                "because it is not a marginal rate tax scale",
+                )
+
+    return combined_tax_scales
